+++ redis/hiredis/test_string.cpp
@@ -161,10 +161,10 @@ static void test(redisContext *c)
 			return;
 		}
 		printf("MGET: type=%d str=%s\n", reply->type, reply->str);
-		for (int i = 0; i < reply->elements; i++)
+		for (size_t i = 0; i < reply->elements; i++)
 		{
 			redisReply *r = reply->element[i];
-			printf("[%d] str=%s\n", i, r->str);
+			printf("[%zu] str=%s\n", i, r->str);
 		}
 		freeReplyObject(reply);
